+++ ruleset/buildins/geo/host_matchers.go
@@ -1,4 +1,4 @@
-package acl
+package geo
 
 import (
 	"net"
